+++ scripts/createCptTpl.js
@@ -1,21 +1,15 @@
 const conf = require('../src/config.json');
-const readline = require('readline');
 const fs = require('fs');
 const path = require('path');
 const inquirer = require('inquirer');
 const copy = require('copy');
+const createPkgDeclare = require('./createPkgDeclare');
 
 let sorts = [...conf.sorts];
 
 let newCpt = {
     version: '1.0.0'
 };
-
-let rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-});
-
 
 function init() {
     inquirer.prompt([
@@ -116,7 +110,7 @@ Vue.${newCpt.type}(${newCpt.name}.name, ${newCpt.name});
 
 export default ${newCpt.name}`;
 
-        const dirPath = path.join(__dirname, `../src/packages/${nameLc}/`);
+        const dirPath = path.join(__dirname, `../src/packages/${nameLc}`);
         const filePath = path.join(dirPath, `index.js`);
         if (!fs.existsSync(dirPath)) {
             fs.mkdirSync(dirPath);
@@ -149,7 +143,7 @@ export default {
     }
 }
 </script>`;
-        const dirPath = path.join(__dirname, `../src/packages/${nameLc}/`);
+        const dirPath = path.join(__dirname, `../src/packages/${nameLc}`);
         const filePath = path.join(dirPath, `${nameLc}.vue`);
         if (!fs.existsSync(dirPath)) {
             fs.mkdirSync(dirPath);
@@ -167,7 +161,7 @@ function createScss() {
         let content = `.nut-${nameLc}{
 
 }`;
-        const dirPath = path.join(__dirname, `../src/packages/${nameLc}/`);
+        const dirPath = path.join(__dirname, `../src/packages/${nameLc}`);
         const filePath = path.join(dirPath, `${nameLc}.scss`);
         if (!fs.existsSync(dirPath)) {
             fs.mkdirSync(dirPath);
@@ -181,13 +175,13 @@ function createScss() {
 
 function createDir() {
     const nameLc = newCpt.name.toLowerCase();
-    const destPath = path.join(__dirname, '../src/packages/' + nameLc + '/');
+    const destPath = path.join(__dirname, '../src/packages/' + nameLc);
     if (!fs.existsSync(destPath)) {
         fs.mkdirSync(destPath);
     }
     copy(path.join(__dirname, './__template__/**.*'), destPath, function (err, file) {
         if (err) {
-            console.log('创建文件夹出错！');
+            console.log('拷贝__template__目录失败！');
         }
         createNew();
     });
@@ -217,6 +211,7 @@ function createNew() {
     }).then(() => {
         return addToPackageJson();
     }).then(() => {
+        createPkgDeclare(newCpt.name);
         console.log('组件模板生成完毕，请开始你的表演~');
         process.exit();
     });